Added live refresh after adding, updating, or deleting from customer database
I lost time due to power outages at home so I decided to make up for it here to demonstrate what I'm capable of under more ideal conditions. This was a very simple fix, all I had to do was use NextJS's useRouter hook to refresh the page after any change to the database. Refreshing the page makes another GET request to my backend and updates the page with the new JSON data.